Added control flow to Lox.
Added jump based opcodes to enable conditional/unconditional jumps.

+++ include/table.h
@@ -16,8 +16,8 @@ typedef struct {
  * @brief Hash table data structure
  */
 typedef struct {
-    size_t count;
-    size_t capacity;
+    uint32_t count;
+    uint32_t capacity;
     Entry *entries;
 } Table;
